+++ uff_wave.cpp
@@ -16,7 +16,7 @@ void Wave::printSelf(std::ostream& os, std::string indent) const
     os << indent << "Origin: ";
     this->m_origin.printSelf(os, indent + UFF_STD_INDENT);
 
-    os << indent << "WaveType: " << this->m_waveType << std::endl;
+    os << indent << "WaveType: " << (int)this->m_waveType << std::endl;
 
     os << indent << "Aperture: ";
     this->m_aperture.printSelf(os, indent + UFF_STD_INDENT);
